thesis_requirement.txt
keras==2.2.4
pydot
pydotplus
pytables
tensorflow==1.12.0
tqdm
pandas==1.0.1
requests
matplotlib
flask
flask-sqlalchemy
flask-login
flask-bcrypt
flask-wtf
psycopg2
libiconv
uwsgi